backend/requirements.txt
astroid==3.2.2
blinker==1.8.2
click==8.1.7
coverage==7.5.1
dill==0.3.8
Flask==3.0.3
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.2
pytest==8.0.2
pytest-cov==4.1.0
pytest-env==1.1.3
python-dotenv==1.0.1
tomlkit==0.12.5
typing_extensions==4.12.0
Werkzeug==3.0.3
psycopg2-binary==2.9.9
